refs: add referent to each_repo_ref_fn
There is no way for callers of the refs api functions that use iterators
to get a hold of a direct value of a symbolic ref before its resolved in
the callback functions, as either each_ref_fn nor each_repo_ref_fn have
an argument for this.

The previous commit started to save the value of the direct reference in
the iterator.

Add an argument to each_repo_ref_fn that gets called with the direct
value of a refernce, and pass the value through from the iterator.

Signed-off-by: John Cai <johncai86@gmail.com>