refs: add referent to each_ref_fn
A past commit allows ref iterators to keep the value of a symref. The
previous commit exposed this value through a parameter added to
each_repo_ref_fn for callers to have access to it.

Add the same parameter to each_ref_fn so that callers to the ref APIs
that use this function as a callback also can have acess to the
unresolved value of a symbolic ref.

Signed-off-by: John Cai <johncai86@gmail.com>